Fix duplicate collectible token ids
